requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
alembic==1.13.1
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1